+++ lib/src/commands/gg_multi_can.dart
@@ -7,8 +7,8 @@
 import 'package:args/command_runner.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:gg_multi/src/commands/can/commit.dart';
-import 'package:gg_multi/src/commands/can/push.dart';
 import 'package:gg_multi/src/commands/can/publish.dart';
+import 'package:gg_multi/src/commands/can/push.dart';
 import 'package:gg_multi/src/commands/can/review.dart';
 
 /// Commands to check if actions can be performed for the current ticket