+++ test/backend/url_parser_test.dart
@@ -4,9 +4,9 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-import 'package:test/test.dart';
-import 'package:gg_multi/src/backend/url_parser.dart';
 import 'package:gg_multi/src/backend/git_platform.dart';
+import 'package:gg_multi/src/backend/url_parser.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('UrlParser', () {